refactored authority_test_utils similarly
