+++ crates/sui-core/Cargo.toml
@@ -81,7 +81,6 @@ sui-swarm-config.workspace = true
 sui-genesis-builder.workspace = true
 sui-json-rpc-types.workspace = true
 sui-macros.workspace = true
-sui-move-build.workspace = true
 sui-network.workspace = true
 sui-protocol-config.workspace = true
 sui-transaction-checks.workspace = true
@@ -93,6 +92,7 @@ zeroize.workspace = true
 nonempty.workspace = true
 
 [dev-dependencies]
+sui-move-build.workspace = true
 clap.workspace = true
 criterion.workspace = true
 expect-test.workspace = true
