+++ crates/sui-core/src/unit_tests/move_package_upgrade_tests.rs
@@ -35,7 +35,7 @@ use crate::authority::move_integration_tests::{
 };
 use crate::authority::test_authority_builder::TestAuthorityBuilder;
 use crate::authority::{
-    authority_test_utils::build_test_modules_with_dep_addr,
+    auth_unit_test_utils::build_test_modules_with_dep_addr,
     authority_tests::execute_programmable_transaction,
     move_integration_tests::build_and_publish_test_package_with_upgrade_cap, AuthorityState,
 };
@@ -556,7 +556,7 @@ async fn test_upgrade_package_invalid_dep_only_upgrade_pre_v68() {
         FileOverlay::Add {
             file_name: "new_friend_module.move",
             contents: r#"
-module base_addr::new_friend_module; 
+module base_addr::new_friend_module;
 public fun friend_call(): u64 { base_addr::base::friend_fun(1) }
         "#,
         },
@@ -595,7 +595,7 @@ async fn test_invalid_dep_only_upgrades() {
         FileOverlay::Add {
             file_name: "new_friend_module.move",
             contents: r#"
-module base_addr::new_friend_module; 
+module base_addr::new_friend_module;
 public fun friend_call(): u64 { base_addr::base::friend_fun(1) }
         "#,
         },
